requirements.txt
matplotlib==3.8.4
mysqlclient==2.2.4
python-dotenv==1.0.1
selenium==4.20.0
